refactor: bottom nav 개선
